Added a BIND inspecting proxy.
If you change passwords and your mail client is constantly presenting the *SAME*
old password, your account will end up getting locked.  This is silly, because
a potential attacker is not gaining any new information by presenting the
*SAME* wrong password.

This proxy recognizes the same bad password is being presented for a DN, and
returns the failed BIND response without telling the proxied server.