Reseting eviction on access should be configurable.
For caching invalid BINDs, resetting the eviction is not a good idea.
E.g. user account is locked for too many different bad passwords.
User uses correct password, but LDAP indicates account is locked
with same error 49.  Now the user's mail clients are all hitting the
cache and keeping the valid passwd in it!