Implemented web service to allow clearing the cache for a particular DN.
Idea is that sometimes a valid password is cached as invalid.

E.g. an account becomes locked.  The user tries to log in, with the
correct password, but the account is locked, so the valid credentials
are cached as invalid.
The account unlocking mechanism could flush the cache for this proxy.